% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_pipeline.R
\name{collapse_nodelist}
\alias{collapse_nodelist}
\title{Recursive helper function for \code{replace_dots}. Collapses nested nodelist back
into a call.}
\usage{
collapse_nodelist(nodelist)
}
\arguments{
\item{nodelist}{A list of calls, with parameters nested, of arbitrary depth}
}
\description{
Recursive helper function for \code{replace_dots}. Collapses nested nodelist back
into a call.
}
\seealso{
\code{\link{replace_dots}}
}
