% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_pipeline.R
\name{get_pipeline}
\alias{get_pipeline}
\title{Utility to grab pipeline expression from wherever available}
\usage{
get_pipeline(pipeline, call)
}
\arguments{
\item{pipeline}{A pipeline. Can be an unquoted or quoted expression, or a
character vector of code. If missing, uses the text highlighted in
RStudio's source editor.}

\item{call}{An expression of the parent call, as captured by \code{\link{match.call}}}
}
\description{
Utility to grab pipeline expression from wherever available
}
