% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_pipeline.R
\name{parse_pipeline}
\alias{parse_pipeline}
\title{Pipeline parser helper function that returns a list of names and calls}
\usage{
parse_pipeline(pipeline, names)
}
\arguments{
\item{pipeline}{A pipeline. Can be an unquoted or quoted expression, or a
character vector of code. If missing, uses the text highlighted in
RStudio's source editor.}

\item{names}{A character vector of names to be used for intermediary
assignment of a length equal to that of the number of top-level calls in
the pipeline.}
}
\value{
A list with \code{names} and \code{calls} elements. The \code{names} element
contains \code{data} and \code{assign} sub-elements.
}
\description{
Pipeline parser helper function that returns a list of names and calls
}
\seealso{
\code{\link{split_pipeline}}, \code{\link{burst_pipes}}
}
