% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_pipeline.R
\name{replace_dots}
\alias{replace_dots}
\title{Recursive helper function that replaces \code{.} within a call with a replacement
expression. Ignores formulas.}
\usage{
replace_dots(call, replacement)
}
\arguments{
\item{call}{A quoted call in which to replace dots}

\item{replacement}{A quoted name with which to replace dots}
}
\value{
A call with the dots replaced
}
\description{
Recursive helper function that replaces \code{.} within a call with a replacement
expression. Ignores formulas.
}
\seealso{
\code{\link{collapse_nodelist}}
}
\keyword{internal}
