% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_pipeline.R
\name{split_pipeline}
\alias{split_pipeline}
\title{Split a pipeline expression into a list of expressions}
\usage{
split_pipeline(pipeline, ...)
}
\arguments{
\item{pipeline}{A pipeline. Can be an unquoted or quoted expression, or a
character vector of code. If missing, uses the text highlighted in
RStudio's source editor.}

\item{...}{Only used to allow assignment with \code{=} at the beginning of a
pipeline.}
}
\value{
A list of unevaluated calls extracted from the pipeline.
}
\description{
\code{split_pipeline} takes an expression of a pipeline and splits it into a list
of unevaluated calls.
}
\details{
\code{split_pipeline} does not split nested pipelines.
}
\examples{
\donttest{
library(magrittr)

split_pipeline(
    quote(x <- 1:5 \%>\% rev \%>\% {. * 2} \%>\% sample(replace = TRUE))
)
}

}
\seealso{
The unexported \href{https://github.com/tidyverse/magrittr/blob/master/R/split_chain.R}{\code{magrittr:::split_chain}},
does not handle assignment, but collects the particular pipes used.
\code{\link{burst_pipes}} invisibly returns a restructured split pipeline.
}
