% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_pipeline.R
\name{debug_pipeline}
\alias{debug_pipeline}
\title{Run a pipeline step-by-step in the debugging browser}
\usage{
debug_pipeline(pipeline, ...)
}
\arguments{
\item{pipeline}{A pipeline. Can be an unquoted or quoted expression, or a
character vector of code. If missing, uses the text highlighted in
RStudio's source editor.}

\item{...}{Only used to allow assignment with \code{=} at the beginning of a
pipeline.}
}
\description{
\code{debug_pipeline} creates a function out of the input pipeline with each line
running an additional step from the pipeline and calls the debugging browser
on the function. Each line run prints to the console plus stores the result
to a variable of the form \code{dot[N]}, where \code{[N]} is the index of the element
in the pipeline, so results can easily be replicated or adjusted in the
debugging console.
}
\details{
\code{debug_pipeline} can also be called interactively in RStudio via the "Debug
pipeline in browser" add-in after highlighting the pipeline to debug in the
source editor.

For a full explanation of the special commands available in the debugging
browser, see \code{\link{browser}}.
}
\examples{
\donttest{
library(magrittr)

debug_pipeline(
    x <- 1:5 \%>\% rev \%>\% {. * 2} \%>\% sample(replace = TRUE)
)
}

}
\seealso{
\code{\link{browser}}, \code{\link[magrittr:debug_pipe]{magrittr::debug_pipe}}, \code{\link{burst_pipes}}
}
